src/alpespartners/bff_partners/requirements.txt
fastapi==0.104.1
strawberry-graphql[fastapi]==0.215.1
sse-starlette==1.6.5
pulsar-client==3.3.0
requests==2.31.0
pydantic-settings==2.0.3
uvicorn==0.24.0